Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,8 +6,8 @@
  **/
 int main(void)
 {
-	int f1 = 0, f2 = 1, i;
-	int next = f1 + f2;
+	long f1 = 0, f2 = 1, i;
+	long next = f1 + f2;
 
 	for (i = 1; i <= 50; ++i)
 	{
